requirements.txt
beautifulsoup4
bs4
python-dotenv

langchain-text-splitters
langchain-community
langgraph
langchain[openai]